Move functions to Utils package
