+++ src/manifest/naming_test.go
@@ -59,14 +59,3 @@ func TestDefaultNaming(t *testing.T) {
 			},
 		))
 }
-
-func TestNamingNormalizeName(t *testing.T) {
-	n := LocalNaming{}
-	assert.Equal(t, "", n.normalizeName(""))
-	assert.Equal(t, "abc", n.normalizeName("abc"))
-	assert.Equal(t, "camel-case", n.normalizeName("CamelCase"))
-	assert.Equal(t, "space-separated", n.normalizeName("   space   separated  "))
-	assert.Equal(t, "abc-def-xyz", n.normalizeName("__abc_def_xyz___"))
-	assert.Equal(t, "abc-dev-xyz", n.normalizeName("--abc-dev-xyz---"))
-	assert.Equal(t, "a-b-cd-e-f-x-y-z", n.normalizeName("a B cd-eF   x_y___z__"))
-}
