+++ src/utils/path.go
@@ -1,4 +1,4 @@
-package manifest
+package utils
 
 import (
 	"fmt"
