+++ src/utils/path_test.go
@@ -1,4 +1,4 @@
-package manifest
+package utils
 
 import (
 	"github.com/stretchr/testify/assert"
